Moved CloseButton.swift to TagListView directory instead of project directory
