+++ TagListView/CloseButton.swift
@@ -9,34 +9,34 @@
 import UIKit
 
 internal class CloseButton: UIButton {
-
+    
     var iconSize: CGFloat = 10
     var lineWidth: CGFloat = 1
-    var lineColor: UIColor = UIColor.whiteColor().colorWithAlphaComponent(0.54)
-
+    var lineColor: UIColor = UIColor.white.withAlphaComponent(0.54)
+    
     weak var tagView: TagView?
-
-    override func drawRect(rect: CGRect) {
+    
+    override func draw(_ rect: CGRect) {
         let path = UIBezierPath()
-
+        
         path.lineWidth = lineWidth
-        path.lineCapStyle = .Round
-
+        path.lineCapStyle = .round
+        
         let iconFrame = CGRect(
             x: (rect.width - iconSize) / 2.0,
             y: (rect.height - iconSize) / 2.0,
             width: iconSize,
             height: iconSize
         )
-
-        path.moveToPoint(iconFrame.origin)
-        path.addLineToPoint(CGPoint(x: iconFrame.maxX, y: iconFrame.maxY))
-        path.moveToPoint(CGPoint(x: iconFrame.maxX, y: iconFrame.minY))
-        path.addLineToPoint(CGPoint(x: iconFrame.minX, y: iconFrame.maxY))
-
+        
+        path.move(to: iconFrame.origin)
+        path.addLine(to: CGPoint(x: iconFrame.maxX, y: iconFrame.maxY))
+        path.move(to: CGPoint(x: iconFrame.maxX, y: iconFrame.minY))
+        path.addLine(to: CGPoint(x: iconFrame.minX, y: iconFrame.maxY))
+        
         lineColor.setStroke()
-
+        
         path.stroke()
     }
-
+    
 }
